requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
numpy==2.1.3
openai==1.54.3
packaging==24.2
pdfminer.six==20240706
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
safetensors==0.4.5
sniffio==1.3.1
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
